Better error handling by catching errors in the evaluator and displaying them to the current output port, moved GetEnvValue to Environment, fixed some access modifiers, create the global environment in the evaluator's constructor. Fix the parser to allow numbers larger than int's range to be parsed, added the rest of the tests for numeric primitives and fixed assertion order.

+++ src/Schemin/Schemin.Parse/Parser.cs
@@ -29,6 +29,7 @@ namespace Schemin.Parse
 {
 	using System;
 	using System.Collections.Generic;
+	using System.Numerics;
 	using Schemin.Tokenize;
 	using Schemin.AST;
 	using Schemin.Evaluate;
@@ -85,7 +86,7 @@ namespace Schemin.Parse
 				case TokenType.Symbol:
 					return ConvertAtom(new ScheminAtom(token.Value));
 				case TokenType.IntegerLiteral:
-					return new ScheminInteger(int.Parse(token.Value));
+					return new ScheminInteger(BigInteger.Parse(token.Value));
 				case TokenType.DecimalLiteral:
 					return new ScheminDecimal(decimal.Parse(token.Value));
 				case TokenType.StringLiteral:
